split GeoprocessingStack from createStack
